Consolidate get_nested_value location to dict_util
