+++ api/src/form_schema/rule_processing/json_rule_util.py
@@ -2,7 +2,7 @@ import logging
 import re
 from typing import Any
 
-import jsonpath_ng
+from src.util.dict_util import get_nested_value
 
 logger = logging.getLogger(__name__)
 
@@ -92,67 +92,6 @@ def get_field_values(data: dict, fields: list[str], path: list[str]) -> list:
             values.append(value)
 
     return values
-
-
-def get_nested_value(data: dict, path: list[str]) -> Any:
-    """Fetch a value from a dictionary based on the nested path
-
-    For example, if you have the following dict:
-        {
-            "path": {
-                "to": {
-                    "some_field": 10
-                },
-                "another_field": "hello"
-            },
-            "array_field": [
-              {
-                "x": 1,
-                "y": "hello"
-              },
-              {
-                "x": 3,
-                "y": "there",
-                "z": "words"
-              }
-            ]
-        }
-
-        Passing in the following paths would give the following values:
-        ["path", "to", "some_field"] -> 10
-        ["path", "another_field"] -> "hello"
-        [] -> Returns the whole dict back
-        ["something", "that", "isn't", "a", "path"] -> None
-
-        Array Cases
-        ["array_field[*]", "x"] -> [2, 3]
-        ["array_field[0]", "x"] -> 2
-        ["array_field[*]", "y"] -> ["hello", "there"]
-        ["array_field[0]", "y"] -> "hello"
-        ["array_field[*]", "z"] -> [None, "there"]
-        ["array_field[0]", "z"] -> None
-        ["array_field[*]"] -> [{"x": 2, "y": "hello"}, {"x": 3, "y": "there", "z": "words"}] # Note this is the same as just ["array_field"] with more steps
-    """
-    # If no path, just return the data
-    if len(path) == 0:
-        return data
-
-    # Use jsonpath_ng to parse the path and
-    # find the data
-    full_path = ".".join(path)
-    expr = jsonpath_ng.parse(full_path)
-    result = expr.find(data)
-
-    # No results, return None, not an empty list
-    if len(result) == 0:
-        return None
-    # One result and the path didn't specify an array (anywhere, not just at end)
-    # then we want to return a single item
-    if len(result) == 1 and "[*]" not in full_path:
-        return result[0].value
-
-    # Otherwise return the list of results
-    return [r.value for r in result]
 
 
 def _populate_nested_value_for_array(
